cli/requirements.txt
requests>=2.31.0
inquirer>=3.1.3
rich>=13.7.0
python-dotenv>=1.0.0
openpyxl>=1.0.0
